fix anonymous_balance and anonymous_coin unit test

+++ crates/sui-framework/packages/sui-framework/sources/crypto/hfe_ops.move
@@ -32,7 +32,7 @@ public native fun hfe_ops_compare_value(
     input_1: u64,
     input_2: u64,
     input_3: u64,
-): u64;
+): u8;
 
 public native fun split_data(
     data: &vector<u8>,
